==> notebooks/03_data_collection_news.ipynb <==
!pip install newsapi-python pandas

from newsapi import NewsApiClient
import pandas as pd
import os
from datetime import datetime

os.makedirs("../data/raw", exist_ok=True)

api_key = "62b37c58bf32400287a1a4b76d4a173c"
newsapi = NewsApiClient(api_key=api_key)

keywords = ["AAPL", "Apple", "TSLA", "Tesla", "MSFT", "Microsoft", "GOOGL", "Google", 
            "AMZN", "Amazon", "S&P500", "S&P 500", "NASDAQ", "IBEX35", "IBEX 35"]
from_date = "2025-07-02"
to_date = datetime.today().strftime('%Y-%m-%d')

# Collect news
all_articles = []

for keyword in keywords:
    print(f"Fetching news for: {keyword}")
    articles = newsapi.get_everything(
        q=keyword,
        from_param=from_date,
        to=to_date,
        language="en",
        sort_by="relevancy",
        page_size=100
    )
    
    for article in articles['articles']:
        all_articles.append([
            article['publishedAt'],
            article['source']['name'],
            article['title'],
            article['description'],
            keyword
        ])

news_df = pd.DataFrame(all_articles, columns=["date", "source", "title", "description", "keyword"])
news_df.to_csv("../data/raw/news_data.csv", index=False)
print("News saved in ../data/raw/news_data.csv")

# Preview
news_df.head()
